Home page template and sidebar
I updated the home page template name to include Revolution Slider,
since the shortcode works to display the slider when the template is
set as the home page.

I’m sure I can make an update to the shortcode so it will display the
slider not matter what page the template is set as, but I’ll leave it
for now.

The badge in the sidebar has been commented out as well.